+++ packages/orq-rc/RELEASES.md
@@ -4079,3 +4079,13 @@ Based on:
 - [typescript v3.14.0-rc.7] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.7 - packages/orq-rc
+
+## 2025-10-06 05:33:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.634.2 (2.721.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.8 - packages/orq-rc